+++ packages/_example/CHANGELOG.md
@@ -8,6 +8,32 @@
 
 * **@forestadmin/agent:** upgraded to 1.0.0-alpha.5
 
+# example [1.0.0-beta.81](https://github.com/ForestAdmin/agent-nodejs/compare/example@1.0.0-beta.80...example@1.0.0-beta.81) (2022-10-05)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/agent:** upgraded to 1.0.0-beta.62
+* **@forestadmin/datasource-dummy:** upgraded to 1.0.0-beta.37
+* **@forestadmin/datasource-sequelize:** upgraded to 1.0.0-beta.45
+* **@forestadmin/datasource-sql:** upgraded to 1.0.0-beta.54
+* **@forestadmin/datasource-toolkit:** upgraded to 1.0.0-beta.35
+* **@forestadmin/datasource-mongoose:** upgraded to 1.0.0-beta.23
+
+# example [1.0.0-beta.80](https://github.com/ForestAdmin/agent-nodejs/compare/example@1.0.0-beta.79...example@1.0.0-beta.80) (2022-10-05)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/agent:** upgraded to 1.0.0-beta.61
+* **@forestadmin/datasource-dummy:** upgraded to 1.0.0-beta.36
+
 # example [1.0.0-alpha.4](https://github.com/ForestAdmin/agent-nodejs/compare/example@1.0.0-alpha.3...example@1.0.0-alpha.4) (2022-09-29)
 
 
@@ -15,6 +41,9 @@
 
 
 ### Dependencies
+
+* **@forestadmin/agent:** upgraded to 1.0.0-alpha.4
+
 
 * **@forestadmin/agent:** upgraded to 1.0.0-alpha.4