+++ packages/agent/CHANGELOG.md
@@ -1,3 +1,24 @@
+# @forestadmin/agent [1.0.0-beta.62](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/agent@1.0.0-beta.61...@forestadmin/agent@1.0.0-beta.62) (2022-10-05)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-customizer:** upgraded to 1.0.0-beta.3
+* **@forestadmin/datasource-toolkit:** upgraded to 1.0.0-beta.35
+
+# @forestadmin/agent [1.0.0-beta.61](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/agent@1.0.0-beta.60...@forestadmin/agent@1.0.0-beta.61) (2022-10-05)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-customizer:** upgraded to 1.0.0-beta.2
+
 # @forestadmin/agent [1.0.0-alpha.5](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/agent@1.0.0-alpha.4...@forestadmin/agent@1.0.0-alpha.5) (2022-10-04)