+++ packages/datasource-customizer/CHANGELOG.md
@@ -1,3 +1,20 @@
+# @forestadmin/datasource-customizer [1.0.0-beta.3](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-customizer@1.0.0-beta.2...@forestadmin/datasource-customizer@1.0.0-beta.3) (2022-10-05)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-toolkit:** upgraded to 1.0.0-beta.35
+
+# @forestadmin/datasource-customizer [1.0.0-beta.2](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-customizer@1.0.0-beta.1...@forestadmin/datasource-customizer@1.0.0-beta.2) (2022-10-05)
+
+
+### Features
+
+* allow including/excluding collection when adding datasources ([#462](https://github.com/ForestAdmin/agent-nodejs/issues/462)) ([cda74d3](https://github.com/ForestAdmin/agent-nodejs/commit/cda74d3d34e8ca04db65f467a9bce41294ebd991))
+
 # @forestadmin/datasource-customizer [1.0.0-alpha.2](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-customizer@1.0.0-alpha.1...@forestadmin/datasource-customizer@1.0.0-alpha.2) (2022-09-28)