+++ packages/datasource-dummy/CHANGELOG.md
@@ -1,3 +1,24 @@
+# @forestadmin/datasource-dummy [1.0.0-beta.37](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-dummy@1.0.0-beta.36...@forestadmin/datasource-dummy@1.0.0-beta.37) (2022-10-05)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-customizer:** upgraded to 1.0.0-beta.3
+* **@forestadmin/datasource-toolkit:** upgraded to 1.0.0-beta.35
+
+# @forestadmin/datasource-dummy [1.0.0-beta.36](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-dummy@1.0.0-beta.35...@forestadmin/datasource-dummy@1.0.0-beta.36) (2022-10-05)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-customizer:** upgraded to 1.0.0-beta.2
+
 # @forestadmin/datasource-dummy [1.0.0-alpha.2](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-dummy@1.0.0-alpha.1...@forestadmin/datasource-dummy@1.0.0-alpha.2) (2022-09-28)